Started Data Point Calculation
