+++ src/features/waves/WaveList.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import { useAppSelector } from "@app/hooks";
-import { selectAllWaves } from "@features/waves/wavesConsts";
 import { WavesListView } from "@components/organisms/WavesListView/WavesListView";
+import { selectAllWaves } from "@features/waves/wavesSelectors";
 
 export const WaveList: FC = () => {
 	const waves = useAppSelector(selectAllWaves);
