+++ src/features/waves/wavesConsts.ts
@@ -1,5 +1,4 @@
-import { IWave, IWavesList, TWavesSlice } from "./wavesInterfaces";
-import { RootState } from "@app/store";
+import { IGetWaveDataPointsProps, IWavesList } from "./wavesInterfaces";
 
 export const initialWavesState: IWavesList = {
 	items: [
@@ -8,46 +7,9 @@ export const initialWavesState: IWavesList = {
 	],
 };
 
-export const addWaveReducer: TWavesSlice["caseReducers"]["addWave"] = (
-	state,
-	action,
-) => {
-	const newWave = action.payload;
-	if (state.items.some((item) => item.id === newWave.id)) {
-		return state;
-	}
-
-	return { ...state, items: [...state.items, newWave] };
-};
-
-export const removeWaveReducer: TWavesSlice["caseReducers"]["removeWave"] = (
-	state,
-	action,
-) => {
-	const removeWave = action.payload;
-
-	return {
-		...state,
-		items: state.items.filter((item) => item.id === removeWave.id),
-	};
-};
-
-//TODO: do i add points calculation here?
-
-export const updateWaveReducer: TWavesSlice["caseReducers"]["updateWave"] = (
-	state,
-	action,
-) => {
-	const updatedWave = action.payload;
-
-	return {
-		...state,
-		items: state.items.map((item) =>
-			item.id === updatedWave.id ? { ...item, ...updatedWave } : item,
-		),
-	};
-};
-
-export const selectAllWaves = (state: RootState): IWave[] => {
-	return state.waves.items;
+export const getCalculatedWaveDataPoints = (
+	props: IGetWaveDataPointsProps,
+): number[] => {
+	const { wave } = props;
+	return [];
 };
